pyproject.toml
[project]
name = "encryption-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=44.0.2",
    "pymongo>=4.12.0",
    "python-dotenv>=1.1.0",
    "streamlit>=1.44.1",
]
